Update GygaxianRandomDungeonGenerator.ps1
A difference in philosophy- I dropped all of the fancier functionality I was adding to the Get-Table5Roll. Instead of including nested rolls on a function that is supposed to represent a specific table, I have started my own new function Get-Room, which will be a curated set of rolls that will complete a module of a dungeon.