Update GygaxianRandomDungeonGenerator.ps1
It took me an embarrassingly long time to figure out an interpretation of the rules that made sense regarding the jewelry table, and it took me an embarrassingly long time to work out the logic, (especially the recursive checks and stacking bonuses here) but I believe I have represented them accurately (at least accurately enough) with the rules as written in the original text. Get-Jewelry is now complete!