Update GygaxianRandomDungeonGenerator.ps1
Minor fixes and had to track down why there was weird behavior. It turns out I'd used the variable "$Jewelry" in two different places and were causing weird errors to be thrown when using "Get-Room" when the treasure result was jewelry and the level (count of the jewelry treasures) was > 1. I have corrected this now.