Update GygaxianRandomDungeonGenerator.ps1
I wrote a general placeholder category output for "Get-Treasure" with the intention of eventually circling back to the tedium of transcribing a bunch of d100 charts for the various magic items that are possible to draw as treasure. That is less appealing in the short term than making progress on the more basic function of this project.