Update GygaxianRandomDungeonGenerator.ps1
Added a comment regarding trapped/hidden treasure:

"This is logic I made up. The DMG indicates that treasure being guarded or hidden is to be rolled "if desired". I think that having one or the other 100% of the time is silly and unintended. I also think that allowing for an escalating probability of dangerous/hidden treasure to increase as the dungeon level increases makes sense, so I have wirrten some logic for what sounds fair to me below. This may need to be amended/revisited in the future."