Update GygaxianRandomDungeonGenerator.ps1
I corrected output to reflect rolls 19 and 20 on table 5C, (the 1 exit from a chamber or room on these results yields the opposite type normally associated with the type- i.e. the room's exit is a passage rather than the normal door, and the chamber's exit is a door rather than a passage). Formatted this output.