Update GygaxianRandomDungeonGenerator.ps1
To the best of my research, caves do not have doors and also do not have corridors, but they can have exits. I have opted to express caves as their own cavey-styled version of a chamber/room by treating it as a room or chamber until its description... then its"type" and "shape" are assigned to "cave" and its exit type is named... "exit" (rather than "door" or "corridor").

I realize this is minutia, but I am at the point where I must concern myself with it.