Update GygaxianRandomDungeonGenerator.ps1
I must be tired. The " x " was the encounter result being called without a "level" parameter, (I guess the function should require it or default to level 1...) so it was representing "(monster) x (number)" when the monster and number values were null. Corrected. The broad strokes are complete on this function.