Update GygaxianRandomDungeonGenerator.ps1
Lots of fixes to "Get-Passage" and its tributary functions. Adding values for consecutive segments of straight hallway (with equal width) was (relatively) simple at first- checking to see if the last two members of the array the function produces has hallway objects, adding them together in the second to last member's value, and then dropping the last member- but this did not work when two segments would become adjacent with another sting segment at the end, (e.g. 30' of 10' wide corridor, then 30' of 10' corridor with a trap at its end). I added a long line of logic to account for this, and another, (now probably superfluous) line to add members 1 and 2, (i.e. 0 and 1) to 2, and then dropping the first member, (the converse of my original pivot). Either way that little issue has been addressed.

I also created a new function that bundles together some sub-rolls that result from "special passages" on table 3B. This is nice, but I didn't even bother trying to incorporate some extraordinarily specific logic hidden in the footnote on that: "Stairs up to gallery... In the former case if a stairway is indicated in or adjacent to the passage it will replace the end stairs 50% (1-10) of the time and supplement 50% (1 1-20) of the time." this is insane detail that is highly situational! Maybe later if I begin saving a function that builds a persistent dungeon map that can be referenced for this check...