Update GygaxianRandomDungeonGenerator.ps1
Transcribed a table from page 219 of the DMG, "Jewelry and Items Typically Bejewelled," which has a footnote explicitly saying that it could be used to identify treasure otherwise simply referred to as "jewelry." I had previously converted a note found earlier in the DMG mentioning a handful of items into a chart for the same purpose before stumbling across this more robust one. I've substituted in this "official" one where my prior one had already been in use.